Update org switch hooks
